third_party/blink/web_tests/external/wpt/css/css-backgrounds/parsing/background-clip-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property background-clip value 'border-box, padding-box, content-box, border-area'
  assert_equals: expected "border-box, padding-box, content-box, border-area" but got "border-box"
[FAIL] Property background-clip value 'border-box' multiple base values
  assert_equals: expected "border-box" but got "border-box, border-box"
[FAIL] Property background-clip value 'border-box, padding-box, content-box' multiple base values
  assert_equals: expected "border-box, padding-box, content-box" but got "border-box, padding-box"
Harness: the test ran to completion.
